server/requirements.txt
auto_mix_prep==0.2.0
Django==5.1.0
django_mongodb_backend==5.1.0b2
langdetect==1.0.9
pytest==7.4.3
python-dotenv==1.1.0
transformers==4.51.3
django-redis==5.2.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
authlib==1.5.2
argon2-cffi==23.1.0
redis==4.5.4
pre-commit==4.2.0
# torch==2.3.1
# tokenizers==0.21.1
# huggingface-hub==0.30.2
# safetensors==0.5.3
# tqdm==4.67.1
# numpy==1.26.4
# sentencepiece>=0.2.0
# optree==1.15.0
